requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
python-dotenv
Flask==3.0.3
Flask-Cors==4.0.1
flask[async]
pydantic==2.8.2
requests
aiohttp
trafilatura==1.12.1
beautifulsoup4==4.12.3
numpy==2.1.0
nltk==3.9.1
textstat==0.7.4
textblob

openai
xgboost-cpu==2.1.1